+++ src/engine/see.rs
@@ -17,8 +17,8 @@ impl SEEContainer {
         for target_piece in PAWN..=KING {
             for attackers in 0..256 {
                 for defenders in 0..256 {
-                    let see = result.evaluate(target_piece as u8, attackers as u8, defenders as u8, &evaluation_parameters);
-                    result.table[target_piece as usize][attackers][defenders] = see;
+                    let see = result.evaluate(target_piece, attackers as u8, defenders as u8, &evaluation_parameters);
+                    result.table[target_piece][attackers][defenders] = see;
                 }
             }
         }
@@ -27,29 +27,36 @@ impl SEEContainer {
     }
 
     /// Gets a result of the static exchange evaluation, based on `attacking_piece`, `target_piece`, `attackers` and `defenders`.
-    pub fn get(&self, attacking_piece: u8, target_piece: u8, attackers: u8, defenders: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
+    pub fn get(&self, attacking_piece: usize, target_piece: usize, attackers: u8, defenders: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
         let attacking_piece_index = self.get_see_piece_index(attacking_piece);
-        let target_piece_index = self.get_see_piece_index(target_piece);
+        let target_piece_index = self.get_see_piece_index(target_piece) as usize;
         let updated_attackers = attackers & !(1 << attacking_piece_index);
 
-        let see = self.table[attacking_piece as usize][defenders as usize][updated_attackers as usize];
+        let see = self.table[attacking_piece][defenders as usize][updated_attackers as usize];
         self.get_piece_value(target_piece_index, evaluation_parameters) - see
     }
 
     /// Evaluates a static exchange evaluation result, based on `target_piece`, `attackers`, `defenders`.
-    fn evaluate(&self, target_piece: u8, attackers: u8, defenders: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
+    fn evaluate(&self, target_piece: usize, attackers: u8, defenders: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
         if attackers == 0 {
             return 0;
         }
 
-        let attacking_piece_index = attackers.get_lsb().bit_scan();
-        let target_piece_index = self.get_see_piece_index(target_piece);
+        let attacking_piece_index = attackers.get_lsb().bit_scan() as usize;
+        let target_piece_index = self.get_see_piece_index(target_piece) as usize;
 
         self.evaluate_internal(attacking_piece_index, target_piece_index, attackers, defenders, evaluation_parameters)
     }
 
     /// Recursive function called by `evaluate` to help evaluate a static exchange evaluation result.
-    fn evaluate_internal(&self, attacking_piece: u8, target_piece: u8, attackers: u8, defenders: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
+    fn evaluate_internal(
+        &self,
+        attacking_piece: usize,
+        target_piece: usize,
+        attackers: u8,
+        defenders: u8,
+        evaluation_parameters: &EvaluationParameters,
+    ) -> i16 {
         if attackers == 0 {
             return 0;
         }
@@ -59,7 +66,7 @@ impl SEEContainer {
         let new_attacking_piece = match defenders {
             0 => 0,
             _ => defenders.get_lsb().bit_scan(),
-        };
+        } as usize;
 
         let see = self.evaluate_internal(new_attacking_piece, attacking_piece, defenders, new_attackers, evaluation_parameters);
         cmp::max(0, target_piece_value - see)
@@ -71,7 +78,7 @@ impl SEEContainer {
     ///  - 2 rooks (index 4-5)
     ///  - 1 queen (index 6)
     ///  - 1 king (index 7)
-    fn get_see_piece_index(&self, piece: u8) -> u8 {
+    fn get_see_piece_index(&self, piece: usize) -> u8 {
         match piece {
             PAWN => 0,
             KNIGHT => 1,
@@ -84,13 +91,13 @@ impl SEEContainer {
     }
 
     /// Gets a piece value based on `piece_index` saved in SEE format (look `get_see_piece_index`).
-    fn get_piece_value(&self, piece_index: u8, evaluation_parameters: &EvaluationParameters) -> i16 {
+    fn get_piece_value(&self, piece_index: usize, evaluation_parameters: &EvaluationParameters) -> i16 {
         match piece_index {
-            0 => evaluation_parameters.piece_value[PAWN as usize] as i16,           // Pawn
-            1 | 2 | 3 => evaluation_parameters.piece_value[BISHOP as usize] as i16, // 3x Knight/bishop
-            4 | 5 => evaluation_parameters.piece_value[ROOK as usize] as i16,       // 2x Rook
-            6 => evaluation_parameters.piece_value[QUEEN as usize] as i16,          // Queen
-            7 => evaluation_parameters.piece_value[KING as usize] as i16,           // King
+            0 => evaluation_parameters.piece_value[PAWN] as i16,           // Pawn
+            1 | 2 | 3 => evaluation_parameters.piece_value[BISHOP] as i16, // 3x Knight/bishop
+            4 | 5 => evaluation_parameters.piece_value[ROOK] as i16,       // 2x Rook
+            6 => evaluation_parameters.piece_value[QUEEN] as i16,          // Queen
+            7 => evaluation_parameters.piece_value[KING] as i16,           // King
             _ => panic!("Invalid value: piece_index={}", piece_index),
         }
     }
