+++ src/evaluation/pst/mod.rs
@@ -12,8 +12,8 @@ pub mod rook;
 /// Evaluates piece-square table value on the `board` and returns score from the white color perspective (more than 0 when advantage, less than 0 when disadvantage).
 /// This evaluator sums all values of the pieces for the specified squares, using incremental counters in `board`.
 pub fn evaluate(board: &Board) -> EvaluationResult {
-    let opening_score = board.pst_scores[WHITE as usize][OPENING as usize] - board.pst_scores[BLACK as usize][OPENING as usize];
-    let ending_score = board.pst_scores[WHITE as usize][ENDING as usize] - board.pst_scores[BLACK as usize][ENDING as usize];
+    let opening_score = board.pst_scores[WHITE][OPENING] - board.pst_scores[BLACK][OPENING];
+    let ending_score = board.pst_scores[WHITE][ENDING] - board.pst_scores[BLACK][ENDING];
 
     EvaluationResult::new(opening_score, ending_score)
 }
@@ -24,17 +24,17 @@ pub fn recalculate_incremental_values(board: &mut Board) {
         for phase in OPENING..=ENDING {
             let mut score = 0;
             for piece_index in PAWN..=KING {
-                let mut pieces = board.pieces[color_index as usize][piece_index as usize];
+                let mut pieces = board.pieces[color_index][piece_index];
                 while pieces != 0 {
                     let square = pieces.get_lsb();
                     let square_index = square.bit_scan();
                     pieces = pieces.pop_lsb();
 
-                    score += board.evaluation_parameters.pst[color_index as usize][piece_index as usize][phase as usize][square_index as usize] as i16;
+                    score += board.evaluation_parameters.pst[color_index][piece_index][phase][square_index as usize] as i16;
                 }
             }
 
-            board.pst_scores[color_index as usize][phase as usize] = score;
+            board.pst_scores[color_index][phase] = score;
         }
     }
 }
