+++ src/state/text/mod.rs
@@ -5,7 +5,7 @@ pub mod moves;
 pub mod pgn;
 
 /// Converts piece `symbol` (p/P, n/N, b/B, r/R, q/Q, k/K) into the corresponding [u8] value. Returns [Err] with the proper error messages when the `symbol` is unknown.
-pub fn symbol_to_piece(symbol: char) -> Result<u8, String> {
+pub fn symbol_to_piece(symbol: char) -> Result<usize, String> {
     match symbol {
         'p' | 'P' => Ok(PAWN),
         'n' | 'N' => Ok(KNIGHT),
@@ -18,7 +18,7 @@ pub fn symbol_to_piece(symbol: char) -> Result<u8, String> {
 }
 
 /// Converts `piece` into the corresponding character (p/P, n/N, b/B, r/R, q/Q, k/K). Returns [Err] with the proper error message when the `piece` is unknown.
-pub fn piece_to_symbol(piece: u8) -> Result<char, String> {
+pub fn piece_to_symbol(piece: usize) -> Result<char, String> {
     match piece {
         PAWN => Ok('P'),
         KNIGHT => Ok('N'),
